Support json enum default value for deserializing unknown values (evolvable enums) (#299)
As raised in #298, when a special request header (`Prefer: include-unknown-enum-members`) is set a returned value that is an EnumType may fail to deserialize (will throw). Microsoft Graph uses a sentinel value `unknownFutureValue` in enums that is an obvious target to use for unknown values. When this enum value is present we can annotate the member with `@JsonEnumDefaultValue` and deserialization will associate unknown values with that member (admittedly with the loss of the actual enumerated value).

This is an important enhancement because Microsoft add new members to existing Graph enumeration types all the time (I see them when I review metadata changes every month). Many enumerations appear to have the `unknownFutureValue` sentinel (231 of 429 EnumTypes across the Graph API have the sentinel now). If a new value is returned (without you specifying the special header) your application can break which is obviously bad. This PR provides protection against that scenario.

From https://learn.microsoft.com/en-us/graph/best-practices-concept#handling-future-members-in-evolvable-enumerations